refactor: 1차 리팩토링
1. 디렉토리 구조변경
