+++ src/components/messages/GameGuideMessage.jsx
